+++ R/globals.R
@@ -1,2 +1,2 @@
 # Declare global variables to avoid R CMD check NOTE for unquoted column names
-utils::globalVariables(c("x", "y", "z", "pet"))
+utils::globalVariables(c("x", "y", "z", "pet", "value"))